Decluttering.
Removed deprecated X-CSP headers.
Removed unused code and styles, and unnecessary comments.
If X mod is not installed, the mod version box will no longer be shown in the mod's page.
Changed minimum width and height values for window size.
Bunch of changes to the main page. (a lot of these are entirely code changes, not visual).